#! Replace old fetch-message with new zfetch based implementation
