+++ src/protocol/mod.rs
@@ -16,17 +16,17 @@ macro_rules! try_multi {
 mod produce;
 mod offset;
 mod metadata;
-mod fetch;
 mod consumer;
 
 mod zreader;
-pub mod zfetch;
+mod fetch_req;
+pub mod fetch;
 
 // ~ re-exports for request/response types defined in submodules
+pub use self::fetch_req::FetchRequest;
 pub use self::produce::{ProduceRequest, ProduceResponse};
 pub use self::offset::{OffsetRequest, OffsetResponse};
 pub use self::metadata::{MetadataRequest, MetadataResponse};
-pub use self::fetch::{FetchRequest, FetchResponse};
 // pub use self::consumer::{ConsumerMetadataRequest, ConsumerMetadataResponse};
 pub use self::consumer::{OffsetFetchRequest, OffsetFetchResponse,
                          OffsetCommitRequest, OffsetCommitResponse};
